fix(eval_policy): update argument handling and improve usage message fix(record): add missing newline for better readability fix(lerobot_record): enhance timing logging for observation and inference

+++ scripts/eval_policy.sh
@@ -1,11 +1,19 @@
 #!/bin/bash
 
-task="Grab the tape and put it in the box."
+policy=$1
+
+if [ -z "$policy" ]; then
+    echo "Usage: $0 <policy> [task]"
+    exit 1
+fi
+
+task="${2:-"pick the tape and place it on the pad."}"
 
 cameras="{ front: {type: opencv, index_or_path: 0, width: 640, height: 480, fps: 30}, \
            wrist: {type: opencv, index_or_path: 2, width: 640, height: 480, fps: 30}}" 
 
-rename_map='{"observation.images.front": "observation.images.camera1", "observation.images.wrist": "observation.images.camera2"}'
+# rename_map='{"observation.images.front": "observation.images.camera1", "observation.images.wrist": "observation.images.camera2"}'
+rename_map='{}'
 
 rm -rf "/home/sorel/.cache/huggingface/lerobot/sorel/eval_so101"
 
@@ -17,10 +25,11 @@ lerobot-record  \
   --robot.id=my_awesome_follower_arm \
   --display_data=true \
   --dataset.repo_id="sorel/eval_so101" \
-  --dataset.push_to_hub=false \
   --dataset.single_task="$task" \
+  --dataset.rename_map="$rename_map" \
+  --dataset.push_to_hub=false \
   --dataset.episode_time_s=3600 \
   --teleop.type=so101_leader \
   --teleop.port=/dev/ttyACM0 \
   --teleop.id=my_awesome_leader_arm \
-  --policy.path="sorel/pi05-so101-record-0121" \
+  --policy.path="$policy"
